% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{create_connection}
\alias{create_connection}
\title{Create a connection to tinycards}
\usage{
create_connection()
}
\value{
list connection with slots useragent, connected, and authenticated
}
\description{
Create a connection to tinycards
}
