% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decks.R
\name{create_deck}
\alias{create_deck}
\title{Create a deck}
\usage{
create_deck(connection, deck)
}
\arguments{
\item{connection}{connection}

\item{deck}{deck object}
}
\value{
boolean success
}
\description{
Create a deck
}
