% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{deck2excel}
\alias{deck2excel}
\title{Convert a deck to an excel file}
\usage{
deck2excel(deck, filename, overwrite = F)
}
\arguments{
\item{deck}{deck}

\item{filename}{where to put the file}

\item{overwrite}{whether to overwrite an existing file}
}
\value{
boolean success
}
\description{
Convert a deck to an excel file
}
