% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decks.R
\name{delete_deck}
\alias{delete_deck}
\title{Delete a deck}
\usage{
delete_deck(connection, deck_id)
}
\arguments{
\item{connection}{connection}

\item{deck_id}{deck id}
}
\value{
boolean success
}
\description{
Delete a deck
}
