% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{excel2deck}
\alias{excel2deck}
\title{Convert an excel file to a deck}
\usage{
excel2deck(filename)
}
\arguments{
\item{filename}{where the file is}
}
\value{
deck
}
\description{
Convert an excel file to a deck
}
