% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decks.R
\name{get_concept}
\alias{get_concept}
\title{Get the Nth concept from a card side}
\usage{
get_concept(card_side, N)
}
\arguments{
\item{card_side}{card side}

\item{N}{N}
}
\value{
concept
}
\description{
Get the Nth concept from a card side
}
