% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decks.R
\name{get_concept_contents}
\alias{get_concept_contents}
\title{Get the contents of a concept}
\usage{
get_concept_contents(concept)
}
\arguments{
\item{concept}{concept}
}
\value{
string
}
\description{
Get the contents of a concept
}
