% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decks.R
\name{get_concept_fact_id}
\alias{get_concept_fact_id}
\title{Get the fact ID of a concept}
\usage{
get_concept_fact_id(concept)
}
\arguments{
\item{concept}{concept}
}
\value{
string
}
\description{
Get the fact ID of a concept
}
