% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decks.R
\name{has_concept}
\alias{has_concept}
\title{Does this card side have an Nth concept?}
\usage{
has_concept(card_side, N)
}
\arguments{
\item{card_side}{card side}

\item{N}{M}
}
\value{
TRUE/FALSE
}
\description{
Does this card side have an Nth concept?
}
