% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decks.R
\name{list_decks}
\alias{list_decks}
\title{List all your current decks}
\usage{
list_decks(connection)
}
\arguments{
\item{connection}{connection}
}
\value{
decklist each item with slots id, name, and image_url
}
\description{
List all your current decks
}
