% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decks.R
\name{load_deck}
\alias{load_deck}
\title{Load a deck}
\usage{
load_deck(connection, deck_id)
}
\arguments{
\item{connection}{connection}

\item{deck_id}{deck id}
}
\value{
list, all about the deck including its cards
}
\description{
Load a deck
}
