% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{login}
\alias{login}
\title{Login to tidycards}
\usage{
login(connection, username, password)
}
\arguments{
\item{connection}{connection}

\item{username}{username}

\item{password}{password}
}
\value{
list connection
}
\description{
Login to tidycards
}
