% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decks.R
\name{update_deck}
\alias{update_deck}
\title{Update a deck}
\usage{
update_deck(connection, deck)
}
\arguments{
\item{connection}{connection}

\item{deck}{deck object}
}
\value{
boolean success
}
\description{
Update a deck
}
